DOOMTOOLS_PATH also needs cygpath call in the shell scripts
Since the DOOMTOOLS_PATH environment variable is also used by
the java code, it needs cygpath translation under cygwin.
This modification update both the path to the jar file that
is fed to java in the shell script and the DOOMTOOLS_PATH
environment variable.

With this change (manually appplied to doomtools) I am able
to run `doomtools --where` and also `doomtool --docs` now works.